requirements.txt
## The following requirements were added by pip freeze:
Django==1.9
djangorestframework==3.9.0
Pillow==5.1.0
pytz==2018.4

